+++ ChatApplication/app/src/main/java/com/example/babarmustafa/chatapplication/MainActivity.java
@@ -5,6 +5,7 @@ import android.content.Intent;
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.ActionMode;
 import android.view.View;
 import android.widget.Button;
@@ -33,6 +34,8 @@ public class MainActivity extends AppCompatActivity {
     private FirebaseAuth auth;
     String usernameforlogin;
     String passwordforlogin;
+    private FirebaseAuth.AuthStateListener mAuthListener;
+    private FirebaseAuth mAuth;
 
 
     @Override
@@ -40,7 +43,9 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        auth = FirebaseAuth.getInstance();
+//        auth = FirebaseAuth.getInstance();
+
+
         s_email = (EditText) findViewById(R.id.e_email);
         s_passs = (EditText) findViewById(R.id.e_pass);
         for_sigin_in = (Button) findViewById(R.id.login_buttobn);
@@ -104,5 +109,44 @@ else if (s_passs.getText().toString().length() == 0) {
                 startActivity(i);
             }
         });
+        mAuthListener = new FirebaseAuth.AuthStateListener() {
+            @Override
+            public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
+                FirebaseUser user = firebaseAuth.getCurrentUser();
+                if (user == null) {
+
+                    // User is signed in
+
+
+//                    Log.d("TAG", "onAuthStateChanged:signed_in:" + user.getUid());
+                } else {
+                    // User is signed out
+                    Intent call = new Intent(MainActivity.this,Chat_Main_View.class);
+                    call.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+                    startActivity(call);
+                    Log.d("TAG", "onAuthStateChanged:signed_out");
+                }
+//                 ...
+            }
+        };
+
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        auth.addAuthStateListener(mAuthListener);
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        auth.removeAuthStateListener(mAuthListener);
+
     }
 }
